refactor: make services more flexible
By passing in all dependencies to each method that requires it rather
than capturing it in the instance state. This should better support more
use cases than the previous structure.

This is a backwards compatible change; the more flexible services are
wrapped by the `DIDCommMessaging` class to behave exactly the same way
as before. To use the new, more flexible services, use
`DIDCommMessagingService` class.

Signed-off-by: Daniel Bluhm <dbluhm@pm.me>